requirements.txt
fastapi==0.78.0
uvicorn==0.18.2
jinja2==3.1.2
python-dotenv==0.20.0
boto3==1.24.10
pydantic==1.8.2
jinja2==3.0.1
